isolate map and add VisitedMapActivity
